requirements.txt
transformers
torch
pandas
scikit-learn
datasets
trl
peft
python-dotenv
tqdm